# Add logic to properly honor naming policy when serializing flag enums