Fixes https://github.com/dotnet/runtime/issues/31622.