@steveharter 2020-05-20
Since naming policies are only one-way the resulting JSON that had a naming policy applied may not be able to be deserialized unless we also apply the naming policy to the CLR enum value names during deserialization (and match them up that way -- like we do with properties).