@layomia 2020-05-20
> Since naming policies are only one-way the resulting JSON that had a naming policy applied may not be able to be deserialized unless we also apply the naming policy to the CLR enum value names during deserialization (and match them up that way -- like we do with properties).

We already apply the naming policy when serializing, just incorrectly in the case of flags.

This change doesn't prohibit any deserialization that was possible prior:

**Serializing `Compressed | IntegrityStream` with camel case**

Before: "compressed, IntegrityStream"
Now: "compressed, integrityStream"

Round tripping is not broken here because we do a case insensitive read with Enum.Parse.

**Serializing `Compressed | IntegrityStream` with a custom snake case policy**

Before: "compressed, _integrity_stream"
After: "compressed, integrity_stream"

Neither of these strings can be read by Enum.Parse, as it expects some ordering of "Compressed, IntegrityStream". So, we are not breaking deserialization here, as it never worked for this scenario.

Ultimately, correctly applying the naming won't affect reading, given the current deserialization implementation. The change is a prerequisite if we want to recognize the policy during deserialization (https://github.com/dotnet/runtime/issues/31619) to enable roundtripping when a naming policy is used (one which does more than return a case-insensitive match).