@steveharter 2020-05-20
>> Round tripping is not broken here because we do a case insensitive read with Enum.Parse.

That works with camel-casing but not with snake-casing...