@layomia 2020-05-20
> > > Round tripping is not broken here because we do a case insensitive read with Enum.Parse.
> 
> That works with camel-casing but not with snake-casing...

Yes, but we _**already**_ honor snake-casing when serializing, and then can't deserialize the output:

```cs
JsonSerializerOptions options = new JsonSerializerOptions { Converters = { new JsonStringEnumConverter(namingPolicy: new SimpleSnakeCasePolicy())} };
FileAttributes val = FileAttributes.IntegrityStream;

string json = JsonSerializer.Serialize(val, options);
Console.WriteLine(json); // "integrity_stream"

JsonSerializer.Deserialize<FileAttributes>(json, options); // JsonException
```

I'll reach out to discuss this offline.